chore: move custom_markers and gps services to map folder 🚚
